requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0.1
django-cors-headers==4.3.1
django-rest-framework==0.1.0
django-storages==1.14.2
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
holidays==0.41
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
pillow==10.2.0
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.17.1
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
xlrd==2.0.1